backend/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6; platform_system == "Windows"
deprecation==2.1.0
flask==3.1.0
functions-framework==3.8.2
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gunicorn==23.0.0; platform_system != "Windows"
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==3.0.2
numpy==2.1.3; python_version == "3.11"
packaging==24.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyjwt==2.10.0
pyparsing==3.2.0; python_version > "3.0"
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
werkzeug==3.1.3
